feat: screens handle empty list better
closes #15

We handle the empty-list case for screens:
 addresses, history, mempool, peers

Displays an animated svg graphic unique to each screen and
some basic text in a card indicating there is no data of that type.

adds reusable EmptyState component for this purpose.

So we no longer show an (ugly) empty table for these screens.